Add a src directory and move logger inside
